Use the 'union' merge strategy for sqitch.plan
Since sqitch.plan is basically an append-only file, this configuration
will ensure that conflicts to this file (i.e., two developers both
adding new changesets to the schema at the same time) will be resolved
by including *both* changes, one after the other, which is the correct
approach.

This is recommended practice by the author of Sqitch (see
https://metacpan.org/module/sqitchtutorial#Emergency).

(Note that this doesn't absolve you from having to think about merging
changes, though; if both developers are modifying the same database
objects, you're still going to have to figure out what the "right"
solution is. For the usual case of two disjoint changes, though, this
patch does The Right Thing.