Add sanity checks to makefile
Verifies that the open source schema directory is a sibling of the
current directory.  Additionally verifies that HEAD of the open source
repo is currently pointing to the correct version, as specified by a
makefile variable.

Yay for home-grown dependency management systems!

(Also added some documentation for the various makefile variables and
targets.)