Makefile tweaks
Instead of relying on having the open source repo checked out in the
proper location, we now perform a local clone and checkout.  This
helps a bit on the developer side, and a lot on the deployment side;
now we just have to specify the version of the enterprise schema
needed.