Update the password_hash_type to have OSC password hashes
This will enable OSC hashed passwords for users to be stored in
an EC install. This is so upgrades from OSC to EC can go smoothly.
The OSC hashed password will be stored in the db with the appropriate
hash type and erchef on first login will convert it to the updated EC
hash type.