Add in missing BEGINs in verify scripts
Without it the verify isn't run in a transaction and a notice is
issued that a transaction isn't being run (and the rollback that was
present would server no purpose).

This isn't that big a deal, since no data was being modified in
the verify scripts, but it should be there for completeness.